Add tests for negative test cases
