+++ src/test/java/bookstore/controller/BookControllerTest.java
@@ -1,16 +1,20 @@
-package bookstore;
+package bookstore.controller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import bookstore.dto.category.CategoryDto;
-import bookstore.dto.category.CategoryRequestDto;
+import bookstore.dto.book.BookDto;
+import bookstore.dto.book.CreateBookRequestDto;
+import bookstore.model.Category;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.math.BigDecimal;
+import java.util.ArrayList;
 import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -25,7 +29,7 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class CategoryControllerTest {
+public class BookControllerTest {
     protected static MockMvc mockMvc;
     @Autowired
     private ObjectMapper objectMapper;
@@ -41,43 +45,47 @@ public class CategoryControllerTest {
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Create a new category")
+    @DisplayName("Create a new book")
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void createCategory_ValidRequestDto_Success() throws Exception {
-        CategoryRequestDto requestDto = createTestRequestDto();
+    void createBook_ValidRequestDto_Success() throws Exception {
+        Category category = createTestCategory();
 
-        CategoryDto expected = createTestCategoryDto();
+        CreateBookRequestDto requestDto = createTestBookRequestDto();
+        requestDto.getCategoryIds().add(category.getId());
+
+        BookDto expected = createTestBookDto();
+        expected.getCategoryIds().add(category.getId());
 
         String jsonRequest = objectMapper.writeValueAsString(requestDto);
 
         MvcResult result = mockMvc.perform(
-                        post("/categories")
+                        post("/books")
                                 .content(jsonRequest)
                                 .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isCreated())
+                .andExpect(status().isOk())
                 .andReturn();
 
-        CategoryDto actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), CategoryDto.class);
-        Assertions.assertNotNull(actual);
-        Assertions.assertNotNull(actual.getId());
-        EqualsBuilder.reflectionEquals(expected, actual, "id");
+        BookDto actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), BookDto.class);
+        assertNotNull(actual);
+        assertNotNull(actual.getId());
+        assertTrue(EqualsBuilder.reflectionEquals(expected, actual, "id"));
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
     @DisplayName("Verify throw exception when request dto is null")
-    void createCategory_CategoryRequestDtoIsNull_ShouldThrowException() throws Exception {
-        CategoryRequestDto requestDto = null;
+    void createBook_BookRequestDtoIsNull_ReturnsBadRequest() throws Exception {
+        CreateBookRequestDto requestDto = null;
 
         String jsonRequest = objectMapper.writeValueAsString(requestDto);
 
         MvcResult result = mockMvc.perform(
-                post("/categories")
-                        .content(jsonRequest)
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/books")
+                                .content(jsonRequest)
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
                 .andReturn();
 
@@ -88,71 +96,75 @@ public class CategoryControllerTest {
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Get all categories")
+    @DisplayName("Get all products")
     @Sql(scripts =
             "classpath:database/books/insert-data.sql",
             executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void getAll_GivenCategoriesInDB_ReturnsAllCategories() throws Exception {
-        MvcResult result = mockMvc.perform(get("/categories")
+    void getAll_GivenBooksInCatalog_ReturnsAllBooks() throws Exception {
+        MvcResult result = mockMvc.perform(get("/books")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
 
-        CategoryDto[] actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), CategoryDto[].class);
-        Assertions.assertNotNull(actual);
-        Assertions.assertEquals(2, actual.length);
+        BookDto[] actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), BookDto[].class);
+        assertNotNull(actual);
+        assertEquals(3, actual.length);
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Get all categories empty list")
+    @DisplayName("Get all books empty list")
     void getAll_EmptyDB_ShouldReturnsEmptyArray() throws Exception {
-        MvcResult result = mockMvc.perform(get("/categories")
+        MvcResult result = mockMvc.perform(get("/books")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
 
-        CategoryDto[] actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), CategoryDto[].class);
+        BookDto[] actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), BookDto[].class);
         assertEquals(0, actual.length);
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Get category by id")
+    @DisplayName("Get book by id")
     @Sql(scripts = "classpath:database/books/insert-data.sql",
             executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void findById_ValidRequestDto_ReturnsCategoryById() throws Exception {
-        CategoryDto expected = createTestCategoryDto();
+    void findById_ValidRequestDto_ReturnsBookById() throws Exception {
+        Category category = createTestCategory();
 
-        MvcResult result = mockMvc.perform(get("/categories/1")
+        BookDto expected = createTestBookDto();
+        expected.getCategoryIds().add(category.getId());
+
+        MvcResult result = mockMvc.perform(get("/books/1")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
 
-        CategoryDto actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), CategoryDto.class);
-        Assertions.assertNotNull(actual);
-        EqualsBuilder.reflectionEquals(expected, actual);
+        BookDto actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), BookDto.class);
+
+        assertNotNull(actual);
+        assertTrue(EqualsBuilder.reflectionEquals(expected, actual, "id"));
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("None existing category id")
+    @DisplayName("None existing book id")
     @Sql(scripts = "classpath:database/books/insert-data.sql",
             executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void findById_WithNoneExistingCategoryId_ShouldThrowException() throws Exception {
-        MvcResult result = mockMvc.perform(get("/categories/5")
+    void findById_WithNoneExistingBookId_ReturnsNotFound() throws Exception {
+        MvcResult result = mockMvc.perform(get("/books/6")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound())
                 .andReturn();
@@ -162,16 +174,30 @@ public class CategoryControllerTest {
         assertEquals(actual, expected);
     }
 
-    private CategoryRequestDto createTestRequestDto() {
-        CategoryRequestDto requestDto = new CategoryRequestDto();
-        requestDto.setName("Fantasy");
+    private Category createTestCategory() {
+        Category category = new Category();
+        category.setId(1L);
+        category.setName("Fantasy");
+        return category;
+    }
+
+    private CreateBookRequestDto createTestBookRequestDto() {
+        CreateBookRequestDto requestDto = new CreateBookRequestDto();
+        requestDto.setTitle("Witcher");
+        requestDto.setAuthor("Sapkovsky");
+        requestDto.setIsbn("1981112314470");
+        requestDto.setPrice(BigDecimal.valueOf(30.75));
+        requestDto.setCategoryIds(new ArrayList<>());
         return requestDto;
     }
 
-    private CategoryDto createTestCategoryDto() {
-        CategoryDto dto = new CategoryDto();
-        dto.setId(1L);
-        dto.setName("Fantasy");
-        return dto;
+    private BookDto createTestBookDto() {
+        BookDto bookDto = new BookDto();
+        bookDto.setTitle("Witcher");
+        bookDto.setAuthor("Sapkovsky");
+        bookDto.setIsbn("1981112314470");
+        bookDto.setPrice(BigDecimal.valueOf(30.75));
+        bookDto.setCategoryIds(new ArrayList<>());
+        return bookDto;
     }
 }
