+++ src/test/java/bookstore/controller/CategoryControllerTest.java
@@ -1,20 +1,17 @@
-package bookstore;
+package bookstore.controller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import bookstore.dto.book.BookDto;
-import bookstore.dto.book.CreateBookRequestDto;
-import bookstore.model.Category;
+import bookstore.dto.category.CategoryDto;
+import bookstore.dto.category.CategoryRequestDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.math.BigDecimal;
-import java.util.ArrayList;
 import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -29,7 +26,7 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class BookControllerTest {
+public class CategoryControllerTest {
     protected static MockMvc mockMvc;
     @Autowired
     private ObjectMapper objectMapper;
@@ -45,47 +42,43 @@ public class BookControllerTest {
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Create a new book")
+    @DisplayName("Create a new category")
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void createBook_ValidRequestDto_Success() throws Exception {
-        Category category = createTestCategory();
+    void createCategory_ValidRequestDto_Success() throws Exception {
+        CategoryRequestDto requestDto = createTestRequestDto();
 
-        CreateBookRequestDto requestDto = createTestBookRequestDto();
-        requestDto.getCategoryIds().add(category.getId());
-
-        BookDto expected = createTestBookDto();
-        expected.getCategoryIds().add(category.getId());
+        CategoryDto expected = createTestCategoryDto();
 
         String jsonRequest = objectMapper.writeValueAsString(requestDto);
 
         MvcResult result = mockMvc.perform(
-                        post("/books")
+                        post("/categories")
                                 .content(jsonRequest)
                                 .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isOk())
+                .andExpect(status().isCreated())
                 .andReturn();
 
-        BookDto actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), BookDto.class);
-        assertNotNull(actual);
-        assertNotNull(actual.getId());
+        CategoryDto actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), CategoryDto.class);
+        Assertions.assertNotNull(actual);
+        Assertions.assertNotNull(actual.getId());
         assertTrue(EqualsBuilder.reflectionEquals(expected, actual, "id"));
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
     @DisplayName("Verify throw exception when request dto is null")
-    void createBook_BookRequestDtoIsNull_ShouldThrowException() throws Exception {
-        CreateBookRequestDto requestDto = null;
+    void createCategory_CategoryRequestDtoIsNull_ReturnsBadRequest() throws Exception {
+        CategoryRequestDto requestDto = null;
 
         String jsonRequest = objectMapper.writeValueAsString(requestDto);
 
         MvcResult result = mockMvc.perform(
-                        post("/books")
-                                .content(jsonRequest)
-                                .contentType(MediaType.APPLICATION_JSON))
+                post("/categories")
+                        .content(jsonRequest)
+                        .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
                 .andReturn();
 
@@ -96,75 +89,71 @@ public class BookControllerTest {
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Get all products")
+    @DisplayName("Get all categories")
     @Sql(scripts =
             "classpath:database/books/insert-data.sql",
             executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void getAll_GivenBooksInCatalog_ReturnsAllBooks() throws Exception {
-        MvcResult result = mockMvc.perform(get("/books")
+    void getAll_GivenCategoriesInDB_ReturnsAllCategories() throws Exception {
+        MvcResult result = mockMvc.perform(get("/categories")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
 
-        BookDto[] actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), BookDto[].class);
-        assertNotNull(actual);
-        assertEquals(3, actual.length);
+        CategoryDto[] actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), CategoryDto[].class);
+        Assertions.assertNotNull(actual);
+        Assertions.assertEquals(2, actual.length);
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Get all books empty list")
+    @DisplayName("Get all categories empty list")
     void getAll_EmptyDB_ShouldReturnsEmptyArray() throws Exception {
-        MvcResult result = mockMvc.perform(get("/books")
+        MvcResult result = mockMvc.perform(get("/categories")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
 
-        BookDto[] actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), BookDto[].class);
+        CategoryDto[] actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), CategoryDto[].class);
         assertEquals(0, actual.length);
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("Get book by id")
+    @DisplayName("Get category by id")
     @Sql(scripts = "classpath:database/books/insert-data.sql",
             executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void findById_ValidRequestDto_ReturnsBookById() throws Exception {
-        Category category = createTestCategory();
+    void findById_ValidRequestDto_ReturnsCategoryById() throws Exception {
+        CategoryDto expected = createTestCategoryDto();
 
-        BookDto expected = createTestBookDto();
-        expected.getCategoryIds().add(category.getId());
-
-        MvcResult result = mockMvc.perform(get("/books/1")
+        MvcResult result = mockMvc.perform(get("/categories/1")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
 
-        BookDto actual = objectMapper.readValue(
-                result.getResponse().getContentAsString(), BookDto.class);
-
-        assertNotNull(actual);
-        assertTrue(EqualsBuilder.reflectionEquals(expected, actual, "id"));
+        CategoryDto actual = objectMapper.readValue(
+                result.getResponse().getContentAsString(), CategoryDto.class);
+        Assertions.assertNotNull(actual);
+        assertTrue(EqualsBuilder.reflectionEquals(expected, actual));
     }
 
     @WithMockUser(username = "admin", roles = {"ADMIN"})
     @Test
-    @DisplayName("None existing book id")
+    @DisplayName("None existing category id")
     @Sql(scripts = "classpath:database/books/insert-data.sql",
             executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
     @Sql(scripts =
             "classpath:database/books/cleanup.sql",
             executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-    void findById_WithNoneExistingBookId_ShouldThrowException() throws Exception {
-        MvcResult result = mockMvc.perform(get("/books/6")
+    void findById_WithNoneExistingCategoryId_ReturnsNotFound() throws Exception {
+        MvcResult result = mockMvc.perform(get("/categories/5")
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isNotFound())
                 .andReturn();
@@ -174,30 +163,16 @@ public class BookControllerTest {
         assertEquals(actual, expected);
     }
 
-    private Category createTestCategory() {
-        Category category = new Category();
-        category.setId(1L);
-        category.setName("Fantasy");
-        return category;
-    }
-
-    private CreateBookRequestDto createTestBookRequestDto() {
-        CreateBookRequestDto requestDto = new CreateBookRequestDto();
-        requestDto.setTitle("Witcher");
-        requestDto.setAuthor("Sapkovsky");
-        requestDto.setIsbn("1981112314470");
-        requestDto.setPrice(BigDecimal.valueOf(30.75));
-        requestDto.setCategoryIds(new ArrayList<>());
+    private CategoryRequestDto createTestRequestDto() {
+        CategoryRequestDto requestDto = new CategoryRequestDto();
+        requestDto.setName("Fantasy");
         return requestDto;
     }
 
-    private BookDto createTestBookDto() {
-        BookDto bookDto = new BookDto();
-        bookDto.setTitle("Witcher");
-        bookDto.setAuthor("Sapkovsky");
-        bookDto.setIsbn("1981112314470");
-        bookDto.setPrice(BigDecimal.valueOf(30.75));
-        bookDto.setCategoryIds(new ArrayList<>());
-        return bookDto;
+    private CategoryDto createTestCategoryDto() {
+        CategoryDto dto = new CategoryDto();
+        dto.setId(1L);
+        dto.setName("Fantasy");
+        return dto;
     }
 }
